+++ game-residentevil5goldedition/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.2.0] - 2026-06-26
+
+- Added: Folder-based installer to correct poorly packaged mods.
+- Added: Direct Launch tool (using re5dx9.exe).
+- Improved: Added modinfo.ini and screenshot.jpg to ignoreDeploy and ignoreConflicts lists to avoid extraneous conflicts.
+
 ## [0.1.1]
 
 - Initial release